Merged PR 5124852: Fix bug in function argument handling
Fix bug in function argument handling

Type names can be bound to symbols in two different situations:
1. In function bodies
2. In domain/resource declaration blocks

In point 2, the only symbol bound to a type name is "this".

In function bodies, we want to output the symbols, because we are
outputing a macro, and secilc will resolve the bindings based on the
(call) call.  In domain and resource declarations, we need to resolve
bindings ourselves.

The previous code always resolved the bindings in av rules.  The new
code takes advantage of two facts about "this":

1. It is always safe to resolve "this"
2. "this" is the only keyword bound to a type in domain and resource
declarations.

This means that we can achieve the desired outcome by always resolving
"this", and never resolving other bound symbols, which is what this
change implements