Set up attributes with inheritance
If type foo inherits from virtual type bar, then bar is a typeattribute in
the generated CIL and foo is a type, and foo should associate with bar.

The current implementation assumes that the parent is virtual and
therefore an attribute.  Future work will be needed to allow inheriting
from non-virtual types