Make HLLErrors::append() consume errors
Most of the time we want to extend a set of errors with new ones.  Using
references to HLLErrors may lead to (re)using the merged error object,
which would then be empty, instead of the new superset of errors.
Taking ownership protects against this potential issue and makes the
calling codes simpler.