Make synthetic resources inherit their direct associated resources
We need to keep track of the hierarchy (i.e. update parent for inherited
associated resources) to be able to properly differenciate synthetic
resources.

However, synthetic names don't contain the whole hierarchy but are kept
short to make them easier to use: <domain-name>-<associated-resource>
This also enables to make sure that a resource can only be associated
once.