Add more constraints to user-defined names and allow to use synthetics
Create a new NameDecl lexer token with more strict constraints to avoid
weird naming. :)  Also allow digits to be used to some extend.

The new Symbol and NameDecl regex enables to avoid LALRPOP token
limitations by making them unambiguous.  Synthetic names are less
constrained but this should not be an issue because they are managed by
Cascade internals.

Use NameDecl to constrain annotation names as well.

Allow '-' in symbol names, which is needed to use synthetic resources.
This makes dash a reserved character for Cascade internal names, which
could be handy in the future.