Merged PR 5303230: Error out on non-virtual inheritance
Error out on non-virtual inheritance

In the long term, we want to support this, but there is a bunch of
remaining work needed for this scenario, and attempting to do this
inheritance now results in invalid CIL being generated.  What we need to
do is make the needed compilation upgrades here and then remove this
check.