Support named args passing with "name=foo" in parser
Nothing is actually done with these yet.  Future work will need to
override the default ordering assumptions.

In our existing docs, we have used this syntax extensively in
annotations, and (to my knowledge) never in function calls.  For
consistency, we should support both.

The near term need is to support this syntax in annotations.  A
follow-on PR will add actual support here.  Allowing these to work in
the parser allows annotations that have no behavior defined yet, but use
the "name=foo" syntax to at least parse correctly.