Add support for named arguments.
This allows specifying arguments by name (name=value syntax).

It also adds preliminary support for default values, but that will be
fully implemented in a follow-up PR.

Here's the story on switching the built-in function structs to owned
types:

Previously, validating an argument didn't need to take data from
definition of the argument to store the validated argument.  This was
good, because for built in functions we create fake arguments which
don't live past the definition of the function, so we can't keep any
reference to them for later (outputting the final policy).

When we add default value support, now we may have some data flow from
the argument definition to the validated argument (namely, the default
value).  In our current policy, we don't define default values for
built-ins, but:

1. I expect that to change in the future.  The resource_transition()
built-in will have a default name field of None/NULL/the empty
string/TBD.
2. AFAIK even without default values in built ins, there's not a clean
way to make the lifetimes work, since it's not possible for the compiler
to prove in the general case that we never use default values for
built-ins.

That means that our built-in functions, which don't live long enough,
flow data to validated arguments.  That means that those validated
arguments need to not store references to the original data.  The main
place that was happening was built-in functions.  Changing all of those
to owned types and creating copies allows the ephemeral function
arguments for built in functions to work with lifetimes.