Implement the core of the @derive annotation
Derive functions from their parents based on annotation.

This changes the output of a number of existing tests slightly.  Our
outputed macros take the type it's a member of it as an argument named
this.  Prior to this commit, rather than use that argument, they simply
used the name.  Since "this" and the name are the same, that's
equivalent without derivation.  Once we derive however, we need them to
say "this", so that we can copy the output and then "this" magically
becomes the child type, because that's the this argument given to the
macro on calls.