Implement "extend" keyword
This keyword allows you to extend a type declared elsewhere.

It is an error to use the extend keyword on an undeclared type.

One minor note is that this moves aliases before associations in the
compilation order.  This is so aliases can be extended.  I don't *think*
there's any reason that associations need to be done first.