Include a backtrace on internal errors
This makes internal errors debuggable.

This relies on the backtrace crate, as opposed to std::backtrace, which
is only available in nightly.  This causes issues with thiserror, which
if it sees a type named "Backtrace" assumes it must be
std::backtrace::Backtrace, and then fails to compile on stable.  Simply
renaming backtrace::Backtrace seems to work around this.