Reject duplicate declarations (same name)
For now, the error message is less than helpful, because it only locates
the second insertion, making them hard to debug.  Further augmentation
of the error handling interface is needed to allow errors to point at
multiple locations.

Additionally, cases where a location isn't available (because the
duplicate is a synthetic insertion) return a generic InternalError.  It
would be sensible to audit/test call sites to ensure that if a duplicate
synthetic insertion is attempted because of user error, that this is
transformed into an appropriate call error based on call site knowledge.