Reject duplicate declarations (same name)
Cases where a location isn't available (because the
duplicate is a synthetic insertion) return a generic InternalError.  It
would be sensible to audit/test call sites to ensure that if a duplicate
synthetic insertion is attempted because of user error, that this is
transformed into an appropriate call error based on call site knowledge.