Handle function name renaming in derivation
If two functions have matching arguments and types, they can be derived
from regardless of argument names, but the child needs a standardized
name.  This name will be each of the arguments, concatenated with "_".

This commit does not handle derived DomTrans rules, as these are
broken in functions anyways, since they store TypeInfos directly in the
DomTransRule struct.