Detect whether we are writing to a terminal and conditionally enable color on error output
Also include a --color argument

I had assumed termcolor did this for us, but according to their
documentation, they do not, and recommend using the atty crate for this
purpose: https://docs.rs/termcolor/latest/termcolor/#detecting-presence-of-a-terminal

The approach in this commit lets casc determine whether or not to use
color, and then error.rs translates that boolean decision into the
appropriate termcolor commands.  Casc makes its decision in turn based
on atty detecting a terminal.

Users may additionally use a --color=WHEN flag, as common for other
utilities.

Reported-by: Matthew Sheets <masheets@linux.microsoft.com>