Add dbus_contexts template
Some info about dbus and SELinux: https://blog.siphos.be/2014/06/d-bus-and-selinux/

SELinux policies ship a dbus_contexts file at
/etc/selinux/[policyname]/contexts/dbus_contexts which provides mappings
from dbus service names to SELinux label.  If a mapping is not defined,
dbus uses the label of the owning process.

The common case is to leave this file empty as with this template and
simply allow the default behavior.

Dbus throws errors and has substantially degraded functionality if this
file is not present at all, so we need to ship a file there.

The current approach doesn't yet provide support for actually specifying
the mappings, but it is full featured enough to make that
straightforward when it is added.  Presumably something like an @dbus()
annotation will be a very straightforward way to specify this in
Cascade, and it can then be plugged in to this machinery.