Convert FuncCall to take a tuple of (Arg, cast)
Previous the argument list is a vector of Arguments.  This change makes
it a tuple of (Argument, CascadeString), where the second element of the
tuple is the name of the cast.

The purpose of this is just to enable storage of the cast in the ast to
prepare for future use of it.