Fix casting "this"
This fixes two bugs around the casting of "this"

1. "this" wasn't being correctly converted into the type it referred to
before being checked whether it was a valid cast.  Therefore it was
always incorrectly considered invalid, since there isn't a valid type
named "this"
2. Once "this" can be cast, conversion of "this" to the correct type in
the outputted CIL needs to happen correctly.  The original casting
implementation stored the cast value as the actual type in the
TypeInstance.  Instead, we need to store the real type there, so that
"this" conversion converts to the original type, not what it is cast to.

Also add tests for the above