Add --package argument to casc
--package takes the output files and packages them up into the contents
to go in /etc/selinux

Currently, we only package the compiled policy.32 and file_contexts.
This commit assumes that the secilc version present generates a
policy.32.  That's an obviously broken assumption that should be fixed
before merging this.  Over the longer term, we should be able to take
the infrastructure added here and add all the other files.

For a policy built without specific system configuration, this names the
final policy "out", which may not be desirable.