Throw internal errors if a range is not available
The previous behavior was to display an error message without a file
pointer.  The intention was that perhaps we would display generic error
messages that didn't correspond to something in the file.  In practice,
this has resulted in displaying messages that should point to something,
but don't because the "source" that contains an error, is actually a
synthetic object generated by the compiler.  In that case, the error
message is confusing, and the bug is actually a Cascade bug that
generated non-compiling synthetic data.  In that situation, we should
clean up and report an error on something the user actually wrote in
their policy, rather than something we generated.

This is a WIP, and unit tests do not yet pass with it in place