Fix bugs in file_context() object class handling
The file_context file_type field need to parse as *both* an SELinux
object class, and also a valid FileType.  Make the FileType names match
the object class names so that this is possible.

Also, update tests, documentation and comments so that this code path is
tested, and the different uses are clearly explained for future
reference.

Reported-by: Matthew Sheets <masheets@microsoft.com>