Mark the last remaining bad error as a TODO for later.
In the error condition demonstrated in virtual_function_association.cas
a function in a synthetically created type should exist but doesn't.  We
should return an error that shows where that synthetical type comes from
and supplies the needed information to the user to correct the error.
That's somewhat non-trivial and outside the scope of this PR.  In order
to expedite the rest of the error handling cleanup, I'd rather leave
that case as an InternalError for now and come back to clean it up
later.

(The previous situation before this series is that that error only
displayed the first line without pointing at a code point).