Run PR failing style checks only on a known good toolchain version
Clippy regularly updates with new checks.  When this happens, we update
our repo to account for this (either fixing the code, or disabling the
checks).  However, this often hits a new PR that happens to be the first
PR to run against the new check.  That PR doesn't necessarily introduce
an issue, but fails anyways.

Split the style checks to run against an older known good version of the
toolchain, which can fail PRs, and also informationally run against the
latest toolchain.  This way, PRs won't block on new clippy checks, but
we'll still be notified of the need to fix them.

Once we fix the latest clippy checks, we can update the known good
version to the new clippy version and continue checking against that
version.

Clippy is the specific example here, but this updates the toolchain for
the full checks step (except building nightly) so that we're running
against a stable known good for all style checking.