Inline format args.
Clippy 1.66 complains about uninlined-format-args

This makes the format strings more readable, although the fact that only
variables can be inlined and not full expressions means that in some
cases we have to use the non-inlined syntax still.  I think the
inconsistency is fine as long as individual format!() calls don't
intermingle inlining and non-inlining.