Adding comment workflow
With previous updates to the clippy workflow
there was attempts to add a comment to the PR
if clippy failed.  This was done under pull_request
action.  This would only work on PRs from the owner
of the repo.  With this comment the comments will
be left under the workflow_run action.