Pass warnings from the library out to casc and print them
The interface between casc and the library is a little more awkward than
I would like because of the tuple.  The reason for that is that
eventually, I want to rework the API entirely where these will be
returning a broader CascadeData struct to make the API calls reentrant.
When that happens, warnings will fit in naturally.  At this point, the
tuple is the cleanest way to prepare for that.