Add a warning on if blocks
The warning indicates that they are not yet implemented.

This commit also adds the piping to pass that warning up to the top
level.  Every time a warning is added, we'll need to add the warning
handling data through the call stack up until someplace that handles
warnings.  Presumably after a few warnings, the coverage of that will
become pretty high.