Include parser support for mls ranges.
This notably omits the "," character which can separate a list of
categories.  The issue is that "," also separates a list of arguments,
so if "," is allowed in contexts, then func(system_u:object_r:type_t:s0:c0,c1)
would be ambiguous as to whether it's one argument (the full context) or
two (the context up until c0 and then a second argument of "c1").
Probably the solution here will be to allow contexts to be optionally
quoted, and require quotes for contexts containing commas, similar to
bash taking strings containing spaces, but that is not yet implemented.