Convert TypeInstances to store references to TypeInfos instead of Cows
This was switched from references to Cows back in May 2022, for reasons
that are unclear.  It doesn't seem to have been necessary and the Owned
case was never used.

The possiblity that they might be Owned causes problems for Contexts,
because Owned TypeInfos could go out of scope with the TypeInstance
containing them does.