Convert Context to store owned TypeInstances instead of TypeInfo references
This is helpful in cases when a TypeInfo can map to multiple actual
values.  A binding to an setype will always be that setype (the type is
the instantiation).  However, non-setypes may have multiple possible
instantiations.  For example, Ports have a port number, and Booleans
have true and false.  This switch enables storing the actual values in
the Context rather than just "this is some port".