Add parser support for optional blocks
We haven't defined exactly how optional policy will work yet, but all
the details on the table involve optional blocks and refpolicy 3 makes
use of optional blocks to retain information from refpolicy 2.

For now, just include all the optional rules in the policy with no
special handling.  This will need to be fixed once we settle on the
details of how optional works.