Support port ranges
There's still some weirdness here.  Two outstanding issues that aren't
addressed by this PR:

1. We ultimately parse ports from strings into u16s, then convert back
to strings, then parse them again.  This is obviously redundant and
weird.  To fix it, we'd need to teach Contexts and TypeInstances about
ports.  Right now they treat ports as just strings, so everything works
fine, but we have to convert back to strings in the middle of
everything.
2. The second parse assumes that a list of ports can be comma-separated,
which makes no sense whatsoever.  It should be unreachable, since a
comma-separated list of ports can't parse.  A list of ports should be a
list.  That would parse, but isn't handled by call_to_portcon_rules().
I don't think port lists are used in the current refpolicy3 draft, so
I'm punting that for now, but it's something we should loop back to.