Support inheritance from non-virtuals
Everything about inheritance should "just work" (knock on wood), except
for the rules.  In other words, type validation, association, derivation
etc don't care if the parent is virtual.

The rules are another matter.  Inheriting from a virtual is implemented
with the virtual being an attribute, so we get all the rules "for free",
by associating the child with the parent and granting access on the
attribute.

In the case of a non-virtual, the parent is an actual type, so we can't
do that.  One possible approach would be to create an attribute for the
parent and grant all parent rules on that instead, but that could be
counter-intuitive for binary analysis.  In order to try to be more
friendly to cross-language compilation and binary analysis, the approach
chosen here was to duplicate all the rules associated with the parent to
apply to the child as well.