Move derive_functions() to before prevalidate_functions()
Originally, derive_functions() was located inside validate_functions()
after the function bodies were validated.  The intent behind this was
that we first ensure that the parents are correct, before generating a
possibly incorrect child.  Unfortuantely, that doesn't actually work
because the parents may call derived functions, and would then fail
validation because the derived function doesn't exist yet.

In order to address this, pull derive_functions() out of
validate_functions() and put it earlier.  With this change the derived
functions get set for prevalidation (so they are castable), and are
available when validating the bodies of all functions.

This requires moving all the rename logic from ValidatedStatements to
Statements, which is the bulk of the diffstat on this commit.

It also has the implication that if we derive from a parent with a
policy bug, we'll be deriving that policy bug into the derived child, so
we ignore validation failures on derived children, unless there are no
failures on other functions, in which case that's a Cascade bug and an
internal error.  This is explained in detail in a comment in
validate_functions().