Ignore clippy manual_flatten errors
1. IMHO this suggestion typically makes the code less readable.  I find
the interior if logic clearer than locating a flatten() in the middle of
a chained collection of iterator calls and recalling the semantics of
flatten() on an iterator.
2. The code suggestions provided by clippy to fix this issue do not
always seem to compile.  I got #159 to compile and make clippy happen,
and the flattening feels readable enough there, but the finally
compiling code does not match clippy's suggestion, which was broken.

(In #159 I appeared to be running into something essentially similar to
https://github.com/rust-lang/rust-clippy/issues/6893, but according to
that issue, it was fixes a few years ago, so I'm not sure what's going
on.  I might dig into it more, but as mentioned in point 1, I'm
skeptical of the value of this check anyways)