Add append_set_map function.
BTreeMap::append() overwrites keys on duplicates.  We have a pattern we
use a few times where we have a map of sets, and instead of overwrite,
we prefer the behavior of inserting the keys from other into the map we
are appending to.

Add a helper function with that behavior