Support empty lists
The main use case for this is that we want to turn on "implicit derive"
in a future PR, where all children that don't use an explicit @derive
annotation get an automatic "@derive(*,*)" behavior added internally, so
that inheritance automatically gives you the functions you expect
without extra boilerplate.  This raises the potential for someone to
want to override this behavior.  One obvious way that jumps to mind to
do that is "@derive([],[])", meaning "derive nothing please".
Alternatives could include the addition of an @noderive annotation,
which might be cleaner, particularly if we want to avoid empty list
support everywhere.  But maybe we want empty lists anyways.