Prioritize the expected type in typeinfo_from_string()
TypeInfo from string is responsible for taking a string and telling us
what it is.  The problem is that some strings might be multiple things.
A notable example is that "bpf" is both a class and a permission in the
capability2 class.  A second example is that a user may name something
in their policy the same thing as a class or permission.

This commit changes typeinfo_from_string() to be aware of what we are
trying to parse in the context and prioritize checks for that, so that
if we're looking for a class we'll try to find a class first, and ignore
other things that may happen to have the same name.

One potential problem remaining for future work is something like:

let foo = bpf;

In the above, we can't tell if "bpf" is a permission or a class just
from the local context.  It would depend how foo is actually used (which
could in theory be ambigious itself.

Probably we should detect this situation and require a type annotation
on the binding, but that is outside the scope of this work I think.

A slightly less malicious case might be something like:

let some_caps = [ audit_read bpf block_suspend ];

In this case, we can locally tell that "bpf" is a permission, because
it's part of a list that needs to all parse to permissions.
Prioritizing the permission check there should be possible, but requires
a little extra plumbing in arg_to_typeinfo_vec, or maybe
Context::insert_from_argument().  That too is left for future work.