Add support for comments in sexps
A comment here is just a regular Atom that starts with a ';'.
sexp_internal surrounds it with newlines, so only that specific atom is
commented out.

Care should be taken if comments are inserted in lists, because they
could change the number of arguments, potentially resulting in invalid
CIL.

nb sexp doesn't natively support comments.  CIL's use of sexp supports
lisp-style comments starting with ";"