Reorder the arguments on resource_transition()
The new signature should better align with expectations by matching the
argument order of type_transition(), and should more closely match
domain_transition().  The original argument order was
selected to better integrate with planned syntactic sugar features that
we decided not to implement, so the arguments for the original order no
longer make sense.