Better type inference for bindings on lists
If we get a list in a binding like:

```
let foo = [ bar baz qux ];
```

We don't know what types to expect when validating.  That could cause
the validation to assume the wrong type, if eg bar is both a domain and
a permission.  So we try our best, and if the list is inconsistent, we
see what types we got and try revalidating the full list against each of
those types.  If we get a match, we use that, otherwise we error.

Currently this assumes only one valid match.  It's possible that we
could have multiple valid matches.  For example:

```
resource read {}
resource write {}

let foo = [ read write ];
```

In this case, we don't know if foo is supposed to be a list of
permissions or a list of resources until we start using it.  As a future
enhancement, we should detect that there are ambiguous possibilities and
prompt the developer to annotate the binding like so:

```
let foo<perm> = [ read write ];
```

But that is future work, not part of this PR