Remove regex dependency
This was recommended by the lalrpop documentation, but is actually a
transitive dependency, not a real one.  With lalrpop 0.19.12 this is now
handled correctly on the lalrpop end.  Lalrpop 0.20 will add support for
a "unicode" feature to handle the whole regex dependency situation even
more cleanly, so we will need to do a little tweaking to our
dependencies at that point.  But this is expected to be good for semver
compatible ^0.19.12