Use the right file when validating default arguments
The previous code had a note about errors being potentially confusing,
and boy was that note accurate.  We were mixing the caller's range with
the declarer's file, and ending up pointing to a nonsense code point in
the declarer's file, and claiming that a random portion of a comment
wasn't a valid type.

The real solution here of course is to store file ids in CascadeStrings
and then every CascadeString will have its file and range together as
one chunk, but that's a pretty major rewrite, that will need to happen
after 0.1