Internal error if recursion checking loops fail to remove
If for some reason, remove() doesn't remove something, then we'll keep
trying to remove it infinitely.  remove() should never fail in this
contexts in the absence of Cascade bugs.

Transform the infinite loop bugs in this case into InternalErrors for
easiest discoverability and debugging in the event that something has
gone wrong.