Report errors if alias names conflict with declared type names
The actual validation and error reporting part here is straightforward.
Most of the complexity of this commit comes from tracking the file the
alias is declared in through to the error site.  That is both complex,
and will likely hurt performance due to file clone()s in a non-error
path.

The good news is that we'll likely get all the performance back, and
reduce code complexity once we implement #219.  When we do that, it's
probably worth revisiting this commit to make sure everything is cleaned
up appropriately and we're not tracking files around for no reason.