Fix bug around function name handling in aliases
The central issue is that when there were no function aliases, we
weren't adding aliases for class aliases.

A second problem is that we weren't skipping aliases for functions that
we weren't deriving.

Something like this:

```
virtual resource a {
	@alias(f2)
	fn f1 (...) { ...}
}

virtual resource b {
	fn f2 (...) { ...}
}

resource c inherits a, b { ... }
```

was resulting in c, getting both the "real" derived f2, and the derived
f1 still having an alias of f2, which gives c two different f2
functions.

In practice however, previous fixes to avoid duplicates would see that
f2 was already defined and skip deriving the real one, which is wrong.