Rewrite association deduplication to handle the resource and domain parts seperately
The motivation here comes from combining similar associations, such as
two different nested associations of the same resource from different
parents, or a nested combined with an annotated association.  We need to
treat associations of the same resource as fundamentally the same, but
having multiple parents, which are the different specific resources
associated.

At the time we are setting up association, there is always one resource
that has already been added to the TypeMap for each association.  For
annotated associations, this is an explicitly declared resource of the
resource name.  For nested ones, the nested declaration declares a
"dom.res" resource.  When we later process annotations, we don't care
whether "res" exists in this case.  So we track these "real" resources,
and build our annotations based on them.